validaciones completadas

+++ js/funcionesCategorias.js
@@ -11,7 +11,7 @@ function consultarCategoriaTodo() {
             $("#TablaResultadoCategorias").append("<tr>");
             $("#TablaResultadoCategorias").append("<th>NOMBRE</th>");
             $("#TablaResultadoCategorias").append("<th>DESCRIPCIÓN</th>");
-            $("#TablaResultadoCategorias").append("<th>CABAÑAS</th>");           
+            $("#TablaResultadoCategorias").append("<th>CABAÑAS</th>");
             $("#TablaResultadoCategorias").append("<th>EDITAR</th>");
             $("#TablaResultadoCategorias").append("<th>ELIMINAR</th>");
             $("#TablaResultadoCategorias").append("</tr>");
@@ -21,11 +21,11 @@ function consultarCategoriaTodo() {
                 $("#TablaResultadoCategorias").append("<td>" + respuesta[i].description + "</td>");
                 let caba = " ";
                 for (j = 0; j < respuesta[i].cabins.length; j++) {
-                    caba += j+1 + ". " + respuesta[i].cabins[j].name + "<br>"
+                    caba += j + 1 + ". " + respuesta[i].cabins[j].name + "<br>"
                 }
-                $("#TablaResultadoCategorias").append("<td>" + caba + "</td>"); 
+                $("#TablaResultadoCategorias").append("<td>" + caba + "</td>");
                 $("#TablaResultadoCategorias").append("<td>" + "<input type='button' value='EDITAR' onclick='traeEditarCategoria(" + respuesta[i].id + ")'>" + "</td>");
-                $("#TablaResultadoCategorias").append("<td>" + "<input type='button' value='ELIMINAR' onclick='eliminarCategoria(" + respuesta[i].id + ")'>" + "</td>");
+                $("#TablaResultadoCategorias").append("<td>" + "<input type='button' value='ELIMINAR' onclick='eliminarCategoria(" + respuesta[i].id + ", " + respuesta[i].cabins.length + ")'>" + "</td>");
                 $("#TablaResultadoCategorias").append("</tr>");
             }
 
@@ -34,27 +34,31 @@ function consultarCategoriaTodo() {
 }
 
 function guardarCategoria() {
-    var datos = {
-        name: $('#nombre').val(),
-        description: $('#descripcion').val()
-    }
-
-    var datosaEnviar = JSON.stringify(datos);
-
-    $.ajax({
-        url: 'http://168.138.144.46:8080/api/Category/save',
-        data: datosaEnviar,
-        type: 'POST',
-        dataType: 'json',
-        contentType: 'application/json',
-        success: function (response) {
-            console.log(response);
-        },
-        complete: function (xhr, status) {
-            alert('Categoría Guardada');
-            limpiarFormulario();
+    if ($('#nombre').val().trim() == "" || $('#descripcion').val().trim() == "") {
+        alert('Hay campos vacíos, por favor verifique');
+    } else {
+        var datos = {
+            name: $('#nombre').val(),
+            description: $('#descripcion').val()
         }
-    });
+
+        var datosaEnviar = JSON.stringify(datos);
+
+        $.ajax({
+            url: 'http://168.138.144.46:8080/api/Category/save',
+            data: datosaEnviar,
+            type: 'POST',
+            dataType: 'json',
+            contentType: 'application/json',
+            success: function (response) {
+                console.log(response);
+            },
+            complete: function (xhr, status) {
+                alert('Categoría Guardada');
+                limpiarFormulario();
+            }
+        });
+    }
 }
 
 function traeEditarCategoria(ide) {
@@ -74,46 +78,56 @@ function traeEditarCategoria(ide) {
 }
 
 function editarCategoria() {
-    var datos = {
-        id: actualizarVar,
-        name: $('#nombre').val(),
-        description: $('#descripcion').val()
-    }
-
-    var datosaEnviar = JSON.stringify(datos);
-
-    $.ajax({
-        url: 'http://168.138.144.46:8080/api/Category/update',
-        data: datosaEnviar,
-        type: 'PUT',
-        dataType: 'json',
-        contentType: 'application/json',
-        success: function (response) {
-            console.log(response);
-        },
-        complete: function (xhr, status) {
-            alert('Categoría Actualizada');
-            consultarCategoriaTodo();
-            limpiarFormulario();
+    if ($('#nombre').val().trim() == "" || $('#descripcion').val().trim() == "") {
+        alert('Hay campos vacíos, por favor verifique');
+    } else {
+        var datos = {
+            id: actualizarVar,
+            name: $('#nombre').val(),
+            description: $('#descripcion').val()
         }
-    });
+
+        var datosaEnviar = JSON.stringify(datos);
+
+        $.ajax({
+            url: 'http://168.138.144.46:8080/api/Category/update',
+            data: datosaEnviar,
+            type: 'PUT',
+            dataType: 'json',
+            contentType: 'application/json',
+            success: function (response) {
+                console.log(response);
+            },
+            complete: function (xhr, status) {
+                alert('Categoría Actualizada');
+                consultarCategoriaTodo();
+                limpiarFormulario();
+            }
+        });
+    }
 }
 
-function eliminarCategoria(ide) {
+function eliminarCategoria(ide, val) {
 
-    $.ajax({
-        url: 'http://168.138.144.46:8080/api/Category/' + ide,
-        type: 'DELETE',
-        dataType: 'json',
-        success: function (response) {
-            console.log(response);
-        },
-        complete: function (xhr, status) {
-            alert('Cabaña Eliminada');
-            consultarCategoriaTodo();
-            limpiarFormulario();
-        }
-    });
+    if (val == 0) {
+        $.ajax({
+            url: 'http://168.138.144.46:8080/api/Category/' + ide,
+            type: 'DELETE',
+            dataType: 'json',
+            success: function (response) {
+                console.log(response);
+            },
+            complete: function (xhr, status) {
+                alert('Categoría Eliminada');
+                consultarCategoriaTodo();
+                limpiarFormulario();
+            }
+        });
+
+    } else {
+        alert('Hay cabañas asociadas, no se puede eliminar categoría');
+    }
+
 }
 
 function limpiarFormulario() {
@@ -122,7 +136,7 @@ function limpiarFormulario() {
     $("#cuartos").val("");
     $("#descripcion").val("");
     $("#categoria").val(1);
-    $("#guardaCab").prop('disabled', false);
-    $("#actualizaCab").prop('disabled', true);
+    $("#guardaCat").prop('disabled', false);
+    $("#actualizaCat").prop('disabled', true);
 }
