+++ js/funcionesClientes.js
@@ -11,7 +11,7 @@ function consultarClienteTodo() {
             $("#TablaResultadoClientes").append("<tr>");
             $("#TablaResultadoClientes").append("<th>NOMBRE</th>");
             $("#TablaResultadoClientes").append("<th>CORREO</th>");
-            $("#TablaResultadoClientes").append("<th>EDAD</th>");          
+            $("#TablaResultadoClientes").append("<th>EDAD</th>");
             $("#TablaResultadoClientes").append("<th>EDITAR</th>");
             $("#TablaResultadoClientes").append("<th>ELIMINAR</th>");
             $("#TablaResultadoClientes").append("</tr>");
@@ -19,9 +19,9 @@ function consultarClienteTodo() {
                 $("#TablaResultadoClientes").append("<tr>");
                 $("#TablaResultadoClientes").append("<td>" + respuesta[i].name + "</td>");
                 $("#TablaResultadoClientes").append("<td>" + respuesta[i].email + "</td>");
-                $("#TablaResultadoClientes").append("<td>" + respuesta[i].age + "</td>"); 
+                $("#TablaResultadoClientes").append("<td>" + respuesta[i].age + "</td>");
                 $("#TablaResultadoClientes").append("<td>" + "<input type='button' value='EDITAR' onclick='traeEditarCliente(" + respuesta[i].idClient + ")'>" + "</td>");
-                $("#TablaResultadoClientes").append("<td>" + "<input type='button' value='ELIMINAR' onclick='eliminarCliente(" + respuesta[i].idClient + ")'>" + "</td>");   
+                $("#TablaResultadoClientes").append("<td>" + "<input type='button' value='ELIMINAR' onclick='eliminarCliente(" + respuesta[i].idClient + ", " + respuesta[i].messages.length + ", " + respuesta[i].reservations.length + ")'>" + "</td>");
                 $("#TablaResultadoClientes").append("</tr>");
             }
 
@@ -30,30 +30,34 @@ function consultarClienteTodo() {
 }
 
 function guardarCliente() {
-    var datos = {
-        name: $("#nombre").val(),
-        email: $("#correo").val(),
-        age: $('#edad').val(),
-        password: $('#contrasena').val()
-        
+    if ($('#nombre').val().trim() == "" || $('#correo').val().trim() == "" || $('#edad').val().trim() == "" || $('#contrasena').val().trim() == "") {
+        alert('Hay campos vacíos, por favor verifique');
+    } else {
+        var datos = {
+            name: $("#nombre").val(),
+            email: $("#correo").val(),
+            age: $('#edad').val(),
+            password: $('#contrasena').val()
+        }
+
+        var datosaEnviar = JSON.stringify(datos);
+
+        $.ajax({
+            url: 'http://168.138.144.46:8080/api/Client/save',
+            data: datosaEnviar,
+            type: 'POST',
+            dataType: 'json',
+            contentType: 'application/json',
+            success: function (response) {
+                console.log(response);
+            },
+            complete: function (xhr, status) {
+                alert('Cliente guardado');
+                limpiarFormulario();
+            }
+        });
     }
 
-    var datosaEnviar = JSON.stringify(datos);
-
-    $.ajax({
-        url: 'http://168.138.144.46:8080/api/Client/save',
-        data: datosaEnviar,
-        type: 'POST',
-        dataType: 'json',
-        contentType: 'application/json',
-        success: function (response) {
-            console.log(response);
-        },
-        complete: function (xhr, status) {
-            alert('Cliente guardado');
-            limpiarFormulario();
-        }
-    });
 }
 
 function traeEditarCliente(ide) {
@@ -75,44 +79,53 @@ function traeEditarCliente(ide) {
 }
 
 function editarCliente() {
-    var datos = {
-        idClient: actualizarVar,
-        name: $('#nombre').val(),
-        email: $('#correo').val(),
-        age: $('#edad').val(),
-        password: $('#contrasena').val()
-    }
-
-    var datosaEnviar = JSON.stringify(datos);
-
-    $.ajax({
-        url: 'http://168.138.144.46:8080/api/Client/update',
-        data: datosaEnviar,
-        type: 'PUT',
-        dataType: 'json',
-        contentType: 'application/json',
-        success: function (response) {
-            console.log(response);
-        },
-        complete: function (xhr, status) {
-            alert('Cliente Actualizado');
-            consultarClienteTodo();
-            limpiarFormulario();
+    if ($('#nombre').val().trim() == "" || $('#correo').val().trim() == "" || $('#edad').val().trim() == "" || $('#contrasena').val().trim() == "") {
+        alert('Hay campos vacíos, por favor verifique');
+    } else {
+        var datos = {
+            idClient: actualizarVar,
+            name: $('#nombre').val(),
+            email: $('#correo').val(),
+            age: $('#edad').val(),
+            password: $('#contrasena').val()
         }
-    });
+
+        var datosaEnviar = JSON.stringify(datos);
+
+        $.ajax({
+            url: 'http://168.138.144.46:8080/api/Client/update',
+            data: datosaEnviar,
+            type: 'PUT',
+            dataType: 'json',
+            contentType: 'application/json',
+            success: function (response) {
+                console.log(response);
+            },
+            complete: function (xhr, status) {
+                alert('Cliente Actualizado');
+                consultarClienteTodo();
+                limpiarFormulario();
+            }
+        });
+    }
 }
 
-function eliminarCliente(ide) {
-    $.ajax({
-        url: 'http://168.138.144.46:8080/api/Client/' + ide,
-        type: 'DELETE',
-        dataType: 'json',
-        success: function (respuesta) {
-            console.log(respuesta)
-            alert('Cliente Eliminado');
-            consultarClienteTodo();
-        }
-    });
+function eliminarCliente(ide, mensaje, reservacion) {
+    if (mensaje == 0 && reservacion == 0) {
+        $.ajax({
+            url: 'http://168.138.144.46:8080/api/Client/' + ide,
+            type: 'DELETE',
+            dataType: 'json',
+            success: function (respuesta) {
+                console.log(respuesta)
+                alert('Cliente Eliminado');
+                consultarClienteTodo();
+            }
+        });
+    } else {
+        alert('Hay mensajes o reservas asociados al cliente, no se puede eliminar, para eliminar, borre mensajes y reservas relacionados');
+    }
+
 }
 
 function limpiarFormulario() {
